+++ common/src/main/java/smartin/miapi/modules/edit_options/material/MaterialViewer.java
@@ -32,6 +32,6 @@ public class MaterialViewer implements EditOption {
     @Environment(EnvType.CLIENT)
     @Override
     public InteractAbleWidget getIconGui(int x, int y, int width, int height, Consumer<EditOption> select, Supplier<EditOption> getSelected) {
-        return new EditOptionIcon(x, y, width, height, select, getSelected, CraftingScreen.BACKGROUND_TEXTURE, 339, 25 + 38 + 56 * 3,  512, 512, "miapi.ui.edit_option.hover.glint", this);
+        return new EditOptionIcon(x, y, width, height, select, getSelected, CraftingScreen.BACKGROUND_TEXTURE, 339, 25 + 38 + 56 * 3,  512, 512, "miapi.ui.edit_option.hover.material", this);
     }
 }