+++ fabric/src/main/java/smartin/miapi/fabric/mixin/LivingEntityMixin.java
@@ -1,22 +1,21 @@
 package smartin.miapi.fabric.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import dev.architectury.event.EventResult;
+import net.minecraft.core.Holder;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.damagesource.DamageSource;
-import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.Attribute;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.Arrow;
 import net.minecraft.world.level.material.FluidState;
-import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import smartin.miapi.attributes.AttributeRegistry;
@@ -25,13 +24,6 @@ import smartin.miapi.mixin.LivingEntityAccessor;
 
 @Mixin(LivingEntity.class)
 public abstract class LivingEntityMixin {
-
-    @Shadow
-    public abstract void stopRiding();
-
-    @Shadow
-    public abstract void forceAddEffect(MobEffectInstance instance, @Nullable Entity entity);
-
     private float storedValue;
     private DamageSource storedDamageSource;
     private MiapiEvents.LivingHurtEvent lastEvent;
@@ -112,15 +104,15 @@ public abstract class LivingEntityMixin {
         return storedDamageSource;
     }
 
-    @ModifyArg(
-            method = "travel",
-            at = @At(
-                    value = "INVOKE",
-                    target = "Lnet/minecraft/world/entity/LivingEntity;moveRelative(FLnet/minecraft/world/phys/Vec3;)V"
-            ),
-            index = 0
+    @ModifyReturnValue(
+            method = "Lnet/minecraft/world/entity/LivingEntity;getAttributeValue(Lnet/minecraft/core/Holder;)D",
+            at = @At(value = "RETURN")
     )
-    private float modifySwimAcceleration(float original) {
+    private double miapi$modifySwimSpeed(double original, Holder<Attribute> attributeHolder) {
+        if(!attributeHolder.is(ResourceLocation.parse("generic.water_movement_efficiency"))){
+            return original;
+        }
+
         LivingEntity self = (LivingEntity) (Object) this;
 
         if (!self.isInWater() || !self.isControlledByLocalInstance()) {
@@ -135,7 +127,6 @@ public abstract class LivingEntityMixin {
         }
 
         double swimSpeed = self.getAttributeValue(AttributeRegistry.SWIM_SPEED);
-
-        return (float)(original * swimSpeed);
+        return original + swimSpeed * 2;
     }
 }